# refactor-results/before-openapi.json
{
  "components": {
    "schemas": {
...
      "ClusterDetail": {
        "additionalProperties": true,
        "properties": {
          "account_concentration": {
            "title": "Account Concentration",
            "type": "number"
          },
          "beneficiary_ids": {
            "items": {
              "type": "string"
            },
            "title": "Beneficiary Ids",
            "type": "array"
          },
          "cluster_id": {
            "title": "Cluster Id",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "related_account_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Account Ids",
            "type": "array"
          },
          "risk_level": {
            "title": "Risk Level",
            "type": "string"
          },
          "risk_score": {
            "title": "Risk Score",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total_disbursed": {
            "title": "Total Disbursed",
            "type": "number"
          }
        },
        "required": [
          "cluster_id",
          "beneficiary_ids",
          "size",
          "risk_score",
          "risk_level",
          "related_account_ids",
          "evidence",
          "account_concentration",
          "total_disbursed"
        ],
        "title": "ClusterDetail",
        "type": "object"
      },
...
      "Evidence": {
        "properties": {
          "beneficiary_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beneficiary Id"
          },
          "contribution": {
            "title": "Contribution",
            "type": "integer"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "indicator": {
            "title": "Indicator",
            "type": "string"
          },
          "related_account_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Account Ids",
            "type": "array"
          },
          "related_beneficiary_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Beneficiary Ids",
            "type": "array"
          },
          "source_record_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Record Ids",
            "type": "array"
          }
        },
        "required": [
          "indicator",
          "contribution",
          "explanation",
          "related_beneficiary_ids",
          "related_account_ids",
          "source_record_ids"
        ],
        "title": "Evidence",
        "type": "object"
      },
...
      "GraphElement": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "title": "GraphElement",
        "type": "object"
      },
      "GraphResponse": {
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphElement"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphElement"
            },
            "title": "Nodes",
            "type": "array"
          },
          "truncated": {
            "default": false,
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "nodes",
          "edges"
        ],
        "title": "GraphResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "Page",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local fictional scholarship audit prototype. Scores are review priorities, not fraud probabilities.",
    "title": "GrantLens",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/clusters": {
      "get": {
        "operationId": "clusters_api_v1_clusters_get",
        "parameters": [
          {
            "in": "query",
            "name": "audit_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audit Id"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "risk_level",
            "required": false,
            "schema": {
              "default": "",
              "pattern": "^(|Low|Medium|High|Critical)$",
              "title": "Risk Level",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 100,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "district",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 100,
              "title": "District",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scheme",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 50,
              "title": "Scheme",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "",
              "pattern": "^(|Needs Review|In Investigation|Verification Requested|Cleared)$",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "risk",
              "pattern": "^(risk|amount|size)$",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "minimum",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100,
              "minimum": 0,
              "title": "Minimum",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "maximum",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 0,
              "title": "Maximum",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clusters"
      }
    },
    "/api/v1/clusters/{cluster_id}": {
      "get": {
        "operationId": "cluster_api_v1_clusters__cluster_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "title": "Cluster Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "audit_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cluster"
      }
    },
    "/api/v1/clusters/{cluster_id}/graph": {
      "get": {
        "operationId": "cluster_graph_api_v1_clusters__cluster_id__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "title": "Cluster Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "audit_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audit Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cluster Graph"
      }
    },
...
    "/api/v1/graph/neighbors": {
      "get": {
        "operationId": "graph_neighbors_api_v1_graph_neighbors_get",
        "parameters": [
          {
            "in": "query",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "audit_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audit Id"
            }
          },
          {
            "in": "query",
            "name": "hops",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 3,
              "minimum": 1,
              "title": "Hops",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Graph Neighbors"
      }
    },
    "/api/v1/graph/path": {
      "get": {
        "operationId": "graph_path_api_v1_graph_path_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": true,
            "schema": {
              "title": "Source",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target",
            "required": true,
            "schema": {
              "title": "Target",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "audit_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audit Id"
            }
          },
          {
            "in": "query",
            "name": "max_hops",
            "required": false,
            "schema": {
              "default": 6,
              "maximum": 6,
              "minimum": 1,
              "title": "Max Hops",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 500,
              "minimum": 2,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Graph Path"
      }
    },
...
    "/api/v1/ui/results": {
      "get": {
        "operationId": "ui_results_api_v1_ui_results_get",
        "parameters": [
          {
            "in": "query",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui Results"
      }
    },
    "/api/v1/ui/summary": {
      "get": {
        "operationId": "ui_summary_api_v1_ui_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scheme",
            "required": false,
            "schema": {
              "default": "",
              "title": "Scheme",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "district",
            "required": false,
            "schema": {
              "default": "",
              "title": "District",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "batch",
            "required": false,
            "schema": {
              "default": "",
              "title": "Batch",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "",
              "title": "Period",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui Summary"
      }
    }
  }
}